Introduce map data structure on the server.
RegionMapFactory now produces a RegionMap object which is passed throughout the stack, instead of a List<Region>. The map can provide optimized lookup methods for getting regions. This allows for faster and simplier code at callsites.

RegionMapFactory and DefaultWorldFactory are both made into singletons that cache their dataset on first load. RegionMapFactory can always return the same RegionMap (as it is deeply immutable) and DefaultWorldFactory can clone the data before returning a copy of it. This prevents having to deserialize JSON data from file on subsequent API calls.

Some inefficient accidentally quadratic sanity checks in the factories are fixed.