Reduce input lag on the client.
When a region is selected on the client for an order, all Regions are re-rendered. useSelectLocation is quite expensive because each region will calculate getActiveBoards and isRetreatTurn.

We can cache these in the WorldContext instead, so each Region can reuse a single result rather than performing redundant work to re-calculate it per-region.

This reduces the re-render cost of Region significantly. Whilst this doesn't solve the input lag issue, it does help reduce it.