Reduce input lag on the client.
When a region is selected on the client for an order, all Regions are re-rendered. This is expensive.

Use an IntersectionObserver on each Map, and don't render the Regions when the map is offscreen. This reduces the rendering workload from all maps, to only onscreen maps.

This results in occasional empty maps when panning since regions will need a moment to load in and not be available right away, however it fixes the input lag issue as only rendering changes for onscreen elements and not all elements keeps the delay to an acceptable level.