Adjudicate performance tweaks (#51)
* Run HasRetreats less often.

World.HasRetreats requires iterating over all units, make it a method since it's not cheap enough to be a property. In Evaluator, run this method once and reuse the result.

* Order.TouchedLocations returns an array.

Since this is a throw away allocation, return an array since that is slightly cheaper than a list. Mark it as a method since it allocates.

* Sped up AdjacencyValidator.EqualsOrIsRelated.

Ultimately this method wants to compare the 4 elements of a location for equality. We can compare the timeline, year and phase up front and skip logic related to determining parent regions until we need it. This provides a nice perf boost whilst also reducing the size of the method.

* Sped up MovementEvaluator.IdentifyRetreats.

Cache a list of all existing retreats outside the loop, to avoid having to reevaluate this for every loop iteration.

* Sped up TouchedOrdersFinder.GetTouchedOrders.

- Change TouchedOrders to a HashSet to make the repeated Contains calls fast.
- Filter only to get final result from depthFirstSearch.TouchedOrders, rather than performing a more expensive set-based Except operation.

* Avoid some List.Contains and List.IndexOf calls.

Use HashSets or direct indexing to avoid these linear scans.

---------

Co-authored-by: Oliveriver <oliveriver1@gmail.com>