Add map rasterisation to improve client responsiveness when zoomed out. (#48)
Currently when zoomed out in the client with many boards visible, scrolling performance is poor. The large number of SVGs visible in the DOM tanks the browser's rendering performance.

To tackle this, introduce a new map-rasterisation mode which caches the map image onto a canvas element. Introduce a policy that when zoomed out sufficiently, we remove the SVGs elements from the DOM and instead display the canvas element.

This greatly improves scrolling performance when zoomed out, as the browser has fewer DOM elements to deal with. The downside is that inputting orders is impossible until zooming back in Since users are likely to zoom in to give the required fidelity for clicking regions to input orders, this seems like a reasonable policy. But others policies are possible.

Code-wise there is a penalty in that we need to create parallel code paths to render elements to a canvas, as well as generating elements for React.