Adjudicate performance tweaks.
- AdjustmentEvaluator creates a lookup for active builds outside the loop.
- Evaluator creates a lookup for boards outside the loop.
- TouchedOrdersFinder creates a lookup for world orders outside the loop.

All of these scenarios avoid a quadratic-like runtime behaviour where every item from one collection is compared with every item from another. We fix them all with a lookup of some kind.

The last scenario is the most impactful as it removes the need to compare every order in the world with every active order. In order to make a valid lookup, we need a key that can be compared for equality. If we realise that EqualsOrIsRelated is comparing the parent regions of a location, we can use that as the key.